test: reorganize tests into unit and integration layers
Split the flat tests/ suite into tests/unit/ (package internals in
isolation) and tests/integration/ (the package wired into a web
framework request pipeline), mirroring the layout used in openapipages.

Pure-unit tests are extracted from the mixed framework files:
- UAMiddleware scope-preservation -> tests/unit/test_middleware.py
- FastAPI install_ua/get_ua helpers -> tests/unit/test_fastapi_contrib.py
- Litestar provide_ua cache reuse -> tests/unit/test_litestar_plugin.py

The shared fixture stays at tests/assets/; relocated files reference it
via ../assets/. The per-file S101 ignore glob is widened to tests/**/*.py
so asserts remain allowed in the new subfolders.
