+++ tests/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Integration tests: the package wired into a web framework request pipeline."""
