+++ tests/integration/test_litestar_plugin.py
@@ -1,4 +1,4 @@
-"""Tests for the Litestar contrib plugin."""
+"""Integration tests for the Litestar contrib plugin."""
 
 from __future__ import annotations
 
@@ -7,12 +7,12 @@ from typing import Any, Dict
 import parametrize_from_file as pff
 import pytest
 from httpx import ASGITransport, AsyncClient
-from litestar import Litestar, Request, get
+from litestar import Litestar, get
 from litestar.di import Provide
 from user_agents.parsers import UserAgent
 
 from asgi_user_agents import UADetails
-from asgi_user_agents.contrib.litestar import UAPlugin, provide_ua
+from asgi_user_agents.contrib.litestar import UAPlugin
 
 
 @get("/")
@@ -49,7 +49,7 @@ app = Litestar(route_handlers=[index], plugins=[UAPlugin()])
 
 
 @pytest.mark.asyncio
-@pff.parametrize(path="assets/test_middleware.json")
+@pff.parametrize(path="../assets/test_middleware.json")
 async def test_user_agent_data(ua_string: str, response_data: dict) -> None:
     """Test that both injected dependencies produce the expected data."""
     async with AsyncClient(
@@ -64,21 +64,6 @@ async def test_user_agent_data(ua_string: str, response_data: dict) -> None:
         assert data["is_bot"] is response_data["is_bot"]
         assert data["is_mobile"] is response_data["is_mobile"]
         assert data["raw_family"] == response_data["browser.family"]
-
-
-def test_provide_ua_reuses_cached_instance_from_scope() -> None:
-    """If `scope['ua']` already holds a UADetails, provide_ua reuses that instance."""
-    cached = UADetails({"type": "http", "headers": [(b"user-agent", b"cached/1.0")]})
-    scope: dict = {
-        "type": "http",
-        "headers": [(b"user-agent", b"different/2.0")],
-        "ua": cached,
-        "path": "/",
-        "method": "GET",
-        "query_string": b"",
-    }
-    request: Request = Request(scope)
-    assert provide_ua(request) is cached
 
 
 @pytest.mark.asyncio
